migrate shared types for menuItemExtension
